Move Header component to components folder;
