+++ src/components/Header.js
@@ -1,21 +1,19 @@
 // @flow
 import React from "react";
 import type { Node } from "react";
-import { StyleSheet, StatusBar, View } from "react-native";
+import { StyleSheet, StatusBar } from "react-native";
 import SafeAreaView from "react-native-safe-area-view";
-import { darkBlueGreyTwoColor } from "../../constants/colors";
+import { darkBlueGreyTwoColor } from "../constants/colors";
 
 type Props = {
   children: Node
 };
 
 const Header = ({ children }: Props) => (
-  <View accessibilityTraits={["header"]} style={styles.header}>
-    <SafeAreaView>
-      <StatusBar barStyle="light-content" animated />
-      {children}
-    </SafeAreaView>
-  </View>
+  <SafeAreaView accessibilityTraits={["header"]} style={styles.header}>
+    <StatusBar barStyle="light-content" animated />
+    {children}
+  </SafeAreaView>
 );
 
 const styles = StyleSheet.create({
