+++ src/screens/EventDetailsScreen/component.js
@@ -2,7 +2,6 @@
 import React, { PureComponent } from "react";
 import { Image, View, StyleSheet, ScrollView } from "react-native";
 import type { NavigationScreenProp } from "react-navigation";
-import Header from "./Header";
 import IconButton from "./IconButton";
 import IconItem from "./IconItem";
 import EventMap from "./EventMap";
@@ -10,6 +9,7 @@ import EventOverview from "./EventOverview";
 import Text from "../../components/Text";
 import Button from "../../components/Button";
 import ContentPadding from "../../components/ContentPadding";
+import Header from "../../components/Header";
 import {
   darkBlueGreyTwoColor,
   lightishGreyColor,
@@ -95,6 +95,11 @@ const renderEventDetails = event =>
 
 class EventDetailsScreen extends PureComponent<Props> {
   static defaultProps = {};
+
+  static navigationOptions = {
+    header: null,
+    tabBarVisible: false
+  };
 
   render() {
     const { event, getAssetUrl } = this.props;
